Refactoring kodu po code review na 4programmers.
%Uporzadkowanie pakietow

#BuyModel
isContains->zmiana nazwy na bardziej jednoznaczna

updateQuantity(String[] amount -> updateQuantity(List<Integer> amount)

#BuyProduct

new BigDecimal(0)->BigDecimal.ZERO

#BookDAO
poprawa literowki w metodzie

#BuyOrderServices
poprawa rzucania bledow
zamiast laczenia stringow ->string.format()

#PictureService 
sciezka do zapisu i odczytu zdjec aktualnie w application.properties

dodanie stosu bledow przy obsludze wyjatkow
 
wyrzucenie zbednego else w add i update 

#PublisherService
delete
poprawa obslugi wyjatkow jesli wydawca nie istnieje

SessionOrderService

zrobienie z BuyModel obiektu sesyjnego wyrzucenie requestow

%Usuniecie z interfejsow DAO public z metod

+++ src/main/java/pl/marcinhawelka/bookswebstore/service/PublisherServiceImpl.java
@@ -5,6 +5,7 @@
  */
 package pl.marcinhawelka.bookswebstore.service;
 
+import pl.marcinhawelka.bookswebstore.service.interfaces.PublisherService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -45,12 +46,17 @@ public class PublisherServiceImpl implements PublisherService {
 
     @Override
     public void deletePublisher(Long id) {
+        Publisher publisher = null;
         try {
-            publisherDAO.delete(id);
+           publisher = publisherDAO.findOne(id);
         } catch (IllegalArgumentException e) {
             System.out.println("Can't delete publisher with id:" + id);
-
+            e.printStackTrace(System.out);
         }
+        if(publisher == null){
+            throw new IllegalArgumentException(String.format("Can't delete publisher with id:%d", id));
+        }
+        publisherDAO.delete(id);
     }
 
     @Override
